requirements.txt
asgiref==3.8.1
Django==5.1.1
django-taggit==6.0.0
psycopg2-binary==2.9.9
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
